+++ xarray/core/dataarray.py
@@ -3481,21 +3481,27 @@ class DataArray(AbstractArray, DataWithCoords):
         return self._from_temp_dataset(ds)
 
     def integrate(
-        self, dim: Union[Hashable, Sequence[Hashable]], datetime_unit: str = None
+        self,
+        coord: Union[Hashable, Sequence[Hashable]] = None,
+        datetime_unit: str = None,
+        *,
+        dim: Union[Hashable, Sequence[Hashable]] = None,
     ) -> "DataArray":
         """ integrate the array with the trapezoidal rule.
 
         .. note::
-            This feature is limited to simple cartesian geometry, i.e. dim
+            This feature is limited to simple cartesian geometry, i.e. coord
             must be one dimensional.
 
         Parameters
         ----------
-        dim : hashable, or sequence of hashable
+        coord : hashable, or sequence of hashable
             Coordinate(s) used for the integration.
         datetime_unit : {"Y", "M", "W", "D", "h", "m", "s", "ms", "us", "ns", \
                          "ps", "fs", "as"}, optional
             Can be used to specify the unit if datetime coordinate is used.
+        dim : hashable, or sequence of hashable, optional
+            Deprecated keyword argument. Please use ``coord`` instead.
 
         Returns
         -------
@@ -3504,6 +3510,8 @@ class DataArray(AbstractArray, DataWithCoords):
         See also
         --------
         numpy.trapz: corresponding numpy function
+        Dataset.integrate
+        DataArray.differentiate
 
         Examples
         --------
@@ -3528,7 +3536,24 @@ class DataArray(AbstractArray, DataWithCoords):
         array([5.4, 6.6, 7.8])
         Dimensions without coordinates: y
         """
-        ds = self._to_temp_dataset().integrate(dim, datetime_unit)
+        if dim is not None:
+            if coord is not None:
+                raise TypeError("cannot specify both 'dim' and 'coord'")
+            warnings.warn(
+                "The `dim` keyword argument to `DataArray.integrate` is "
+                "being renamed to `coord`, for consistency with "
+                "`DataArray.differentiate` and `Dataset.integrate`.",
+                FutureWarning,
+                stacklevel=2,
+            )
+            coord = dim
+
+        if coord is None:
+            raise TypeError(
+                "integrate() missing 1 required positional argument: 'coord'"
+            )
+
+        ds = self._to_temp_dataset().integrate(coord, datetime_unit)
         return self._from_temp_dataset(ds)
 
     def unify_chunks(self) -> "DataArray":